requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.28
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
Flask-WTF==1.2.1
alpaca-py==0.13.3
pytz==2024.1

# Testing dependencies
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1
freezegun==1.4.0 